requirements.txt
pdfplumber==0.11.4  # Para extracción de PDFs estructurados
pytesseract==0.3.10  # Para OCR
Pillow==10.4.0  # Requerido por pytesseract para imágenes
pdf2image==1.17.1  # Para convertir PDFs a imágenes
PyPDF2==4.2.0  # Para detectar texto en PDFs
opencv-python==4.10.1.26  # Para detección de regiones y preprocesamiento avanzado
scikit-image==0.24.1  # Para operaciones morfológicas y análisis de imagen
pandas==2.2.3  # Procesamiento de datos
numpy==2.1.1  # Para cálculos
sqlalchemy==2.0.35  # ORM para BD
psycopg2-binary==2.9.9  # Driver para PostgreSQL
APScheduler==3.10.4  # Scheduler para tareas periódicas
yagmail==0.15.293  # Para emails simples
openpyxl==3.1.5  # Para generar Excels
fastapi==0.115.0  # Para API y dashboard
uvicorn==0.30.6  # Server para FastAPI
transformers==4.44.2  # Hugging Face para IA/NLP
torch==2.4.1  # Requerido por transformers (elige versión CPU si no tienes GPU)
python-dotenv==1.0.1  # Para cargar .env
pytest==8.3.3  # Para tests